Gui: split selection observer implementation
